Make draft-store a ports-and-adapters unit test
