+++ packages/app/src/stores/draft-store.migration.test.ts
@@ -1,7 +1,13 @@
-import "@/test/window-local-storage";
-import { beforeEach, describe, expect, it } from "vitest";
+import { describe, expect, it } from "vitest";
 import type { ComposerAttachment } from "@/attachments/types";
-import { __draftStoreTestUtils, useDraftStore } from "./draft-store";
+import {
+  isAttachmentMetadata,
+  migratePersistedState,
+  type MigrateLegacyImages,
+} from "./draft-store.pure";
+
+const passThroughMigrateLegacyImages: MigrateLegacyImages = async (images) =>
+  images.filter(isAttachmentMetadata);
 
 function workspaceReviewAttachment(): Extract<ComposerAttachment, { kind: "review" }> {
   return {
@@ -44,34 +50,33 @@ function workspaceReviewAttachment(): Extract<ComposerAttachment, { kind: "revie
 }
 
 describe("draft-store migration", () => {
-  beforeEach(() => {
-    useDraftStore.setState({ drafts: {}, createModalDraft: null });
-  });
-
   it("normalizes legacy image metadata into image attachments and strips persisted preview URLs", async () => {
-    const migrated = await __draftStoreTestUtils.migratePersistedState({
-      drafts: {
-        "agent:server:agent": {
-          input: {
-            text: "hello",
-            images: [
-              {
-                id: "att-1",
-                mimeType: "image/png",
-                storageType: "desktop-file",
-                storageKey: "/tmp/att-1.png",
-                createdAt: 1700000000000,
-                previewUri: "asset://should-not-persist",
-              },
-            ],
+    const migrated = await migratePersistedState(
+      {
+        drafts: {
+          "agent:server:agent": {
+            input: {
+              text: "hello",
+              images: [
+                {
+                  id: "att-1",
+                  mimeType: "image/png",
+                  storageType: "desktop-file",
+                  storageKey: "/tmp/att-1.png",
+                  createdAt: 1700000000000,
+                  previewUri: "asset://should-not-persist",
+                },
+              ],
+            },
+            lifecycle: "active",
+            updatedAt: 1700000000001,
+            version: 1,
           },
-          lifecycle: "active",
-          updatedAt: 1700000000001,
-          version: 1,
         },
+        createModalDraft: null,
       },
-      createModalDraft: null,
-    });
+      { migrateLegacyImages: passThroughMigrateLegacyImages, nowMs: 1700000000002 },
+    );
 
     expect(migrated.drafts["agent:server:agent"]?.input).toEqual({
       text: "hello",
@@ -118,8 +123,9 @@ describe("draft-store migration", () => {
       createModalDraft: null,
     };
 
-    const once = await __draftStoreTestUtils.migratePersistedState(original);
-    const twice = await __draftStoreTestUtils.migratePersistedState(once);
+    const ports = { migrateLegacyImages: passThroughMigrateLegacyImages, nowMs: 1700000000002 };
+    const once = await migratePersistedState(original, ports);
+    const twice = await migratePersistedState(once, ports);
 
     expect(twice).toEqual(once);
     expect(twice.drafts["agent:server:agent"]?.input).toEqual({
@@ -140,20 +146,23 @@ describe("draft-store migration", () => {
   });
 
   it("rejects workspace review attachments from migrated draft attachments", async () => {
-    const migrated = await __draftStoreTestUtils.migratePersistedState({
-      drafts: {
-        "agent:server:agent": {
-          input: {
-            text: "hello",
-            attachments: [workspaceReviewAttachment()],
+    const migrated = await migratePersistedState(
+      {
+        drafts: {
+          "agent:server:agent": {
+            input: {
+              text: "hello",
+              attachments: [workspaceReviewAttachment()],
+            },
+            lifecycle: "active",
+            updatedAt: 1700000000001,
+            version: 2,
           },
-          lifecycle: "active",
-          updatedAt: 1700000000001,
-          version: 2,
         },
+        createModalDraft: null,
       },
-      createModalDraft: null,
-    });
+      { migrateLegacyImages: passThroughMigrateLegacyImages, nowMs: 1700000000002 },
+    );
 
     expect(migrated.drafts["agent:server:agent"]?.input.attachments).toEqual([]);
   });
